Plaid/requirements.txt
Flask==2.1.0
plaid-python==13.0.0
python-dotenv==1.0.0
